+++ frontend/src/components/Navigation/ProfileButton.jsx
@@ -1,0 +1,33 @@
+import { useState } from 'react';
+import { useDispatch } from 'react-redux';
+import { FaUserCircle } from 'react-icons/fa';
+import * as sessionActions from '../../store/session';
+
+function ProfileButton({ user }) {
+  const dispatch = useDispatch();
+
+  const logout = (e) => {
+    e.preventDefault();
+    dispatch(sessionActions.logout());
+  };
+
+  return (
+    <>
+      <button>
+        <FaUserCircle />
+      </button>
+      <ul className="profile-dropdown">
+        <li>{user.username}</li>
+        <li>
+          {user.firstName} {user.lastName}
+        </li>
+        <li>{user.email}</li>
+        <li>
+          <button onClick={logout}>Log Out</button>
+        </li>
+      </ul>
+    </>
+  );
+}
+
+export default ProfileButton;
